+++ packages/runtime/src/mock/tx.ts
@@ -1,19 +1,21 @@
+import { types } from "@algo-builder/web";
 import { SuggestedParams } from "algosdk";
 
 import { ALGORAND_MIN_TX_FEE } from "../lib/constants";
-import { TxParams } from "../types";
 
 const GENESIS_ID = 'testnet-v1.0';
 // testnet-v1.0 hash
 const GENESIS_HASH = 'SGO1GKSzyE7IEPItTxCByw9x8FmnrCDexi9/cOUJOiI=';
 
 export function mockSuggestedParams (
-  payFlags: TxParams, round: number): SuggestedParams {
+  payFlags: types.TxParams, round: number): SuggestedParams {
   const s = {} as SuggestedParams;
 
-  s.flatFee = payFlags.totalFee !== undefined;
+  if (payFlags.flatFee === undefined) {
+    if (payFlags.totalFee !== undefined) s.flatFee = true;
+    else s.flatFee = false;
+  }
   s.fee = payFlags.totalFee ?? payFlags.feePerByte ?? ALGORAND_MIN_TX_FEE;
-  if (s.flatFee) s.fee = Math.max(s.fee, ALGORAND_MIN_TX_FEE);
 
   // https://developer.algorand.org/docs/features/transactions/#setting-first-and-last-valid
   s.firstRound = payFlags.firstValid ?? (round - 1);